Extract media admin page models
